validate: refactor git-validate
Do not define GIT_CHECK_EXCLUDE in two places and just define it once
and make the action use the Makefile.
In addtion we don't want to lint the full set of commits, this seems to
cause side effects with merge commits as it lints commits that may no
actual be in the PR and thus it can fail.

Using "github.event.pull_request.base.sha" is the right base commit we
care about and makes the validation work. Thanks to Miloslav for finding
that value.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit f25011560dcd9795be6016fcde2b0e393f8b73ec)
Signed-off-by: Paul Holzinger <pholzing@redhat.com>